+++ example/a.ts
@@ -1,7 +1,5 @@
 import { Elysia, t } from '../src'
 
-const app = new Elysia().post('/', ({ body: { file } }) => file, {
-	body: t.Object({
-		file: t.String()
-	})
-})
+const app = new Elysia()
+	.get('/', () => 'A')
+	.listen(3000)